+++ cm-mlops/script/generate-mlperf-inference-submission/customize.py
@@ -59,7 +59,7 @@ def generate_submission(i):
         system_meta['division'] = env['CM_MLPERF_SUBMISSION_DIVISION']
 
     if 'CM_MLPERF_SUBMISSION_CATEGORY' in env:
-        system_meta['system_type'] = env['CM_MLPERF_SUBMISSION_CATEGORY']
+        system_meta['system_type'] = env['CM_MLPERF_SUBMISSION_CATEGORY'].replace("-", ",")
 
     duplicate= (env.get('CM_MLPERF_DUPLICATE_SCENARIO_RESULTS', 'no') in ["yes", "True"])